Updated text component
